Fix pointer cursor on descendants of primary buttons
The cursor in primary buttons is shown as a pointer (if the button is
not disabled) to convey that it can be interacted with. However, the
descendants of the button may not inherit the cursor from it (for
example, a "strong" element would, but a "span" element would not),
which causes a pointer cursor to be shown on some areas of the button
and a different one to be shown on other areas. To prevent that now all
the descendants of a primary button use a pointer cursor.

Moreover, although ".primary" is typically applied on buttons it can be
applied to other elements too. Due to that the pointer cursor might be
shown on elements in which it did not make much sense (like text areas),
so the pointer cursor is now limited to primary buttons only.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>